// apps/web/src/shared/api/generated/model/listsControllerSearchSort.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type ListsControllerSearchSort =
  (typeof ListsControllerSearchSort)[keyof typeof ListsControllerSearchSort];

export const ListsControllerSearchSort = {
  updated_desc: "updated_desc",
  created_desc: "created_desc",
  created_asc: "created_asc",
  title_asc: "title_asc",
  title_desc: "title_desc",
  books_count_desc: "books_count_desc",
  books_count_asc: "books_count_asc",
} as const;